Major Advancements Co-authored-by: heetmehta18 <heetmehta18125@gmail.com>
