+++ README.md
@@ -37,4 +37,3 @@ Feel free to contribute to this project by:
 
 - Improving planet details or adding new celestial bodies.
 - Enhancing the user interface or animations.
-- Suggesting new features like a search bar or 3D effects.
