+++ backend/Makefile
@@ -1,0 +1,5 @@
+all:
+	gcc simulation.c -o simulation -lm
+
+run: all
+	./simulation
