--- solar_system_project/assets/css/styles.css
@@ -1,100 +1,0 @@
-body {
-    margin: 0;
-    background: #000016;
-    overflow: hidden;
-    font-family: Arial, sans-serif;
-}
-
-.solar-system {
-    width: 100vw;
-    height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: relative;
-}
-
-.sun {
-    width: 80px;
-    height: 80px;
-    background: radial-gradient(#ffd700, #ff4500);
-    border-radius: 50%;
-    position: absolute;
-    box-shadow: 0 0 60px #ff4500;
-    cursor: pointer;
-    z-index: 100;
-}
-
-.orbit {
-    border: 1px solid rgba(255, 255, 255, 0.1);
-    border-radius: 50%;
-    position: absolute;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.planet {
-    position: absolute;
-    border-radius: 50%;
-    cursor: pointer;
-    transition: transform 0.3s;
-    z-index: 10;
-}
-
-.planet:hover {
-    transform: scale(1.5);
-}
-
-.planet-info {
-    display: none;
-    position: fixed;
-    top: 20px;
-    right: 20px;
-    background: rgba(0, 0, 0, 0.9);
-    color: white;
-    padding: 20px;
-    border-radius: 10px;
-    width: 300px;
-    box-shadow: 0 0 20px rgba(255, 255, 255, 0.1);
-    z-index: 1000;
-}
-
-.planet-info h2 {
-    color: #ffd700;
-    margin-top: 0;
-}
-
-.planet-info table {
-    width: 100%;
-    margin-top: 10px;
-}
-
-.planet-info td {
-    padding: 5px 0;
-}
-
-.planet-info td:first-child {
-    color: #aaa;
-    width: 40%;
-}
-
-/* Stars styles */
-.stars {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100vw;
-    height: 100vh;
-    pointer-events: none;
-    z-index: 1;
-}
-
-.star {
-    position: absolute;
-    width: 2px;
-    height: 2px;
-    background-color: white;
-    border-radius: 50%;
-    opacity: 0.8;
-}
