--- solar_system_project/backend/Makefile
@@ -1,5 +1,0 @@
-all:
-	gcc simulation.c -o simulation -lm
-
-run: all
-	./simulation
