--- solar_system_project/frontend/index.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Solar System</title>
-    <link rel="stylesheet" href="../assets/css/styles.css">
-</head>
-<body>
-
-    <div class="solar-system">
-        <div class="sun" onclick="showPlanetInfo('sun')"></div>
-        <div class="planet-info" id="planetInfo"></div>
-    </div>
-
-    <script src="../assets/js/script.js"></script>
-    
-</body>
-</html>
-
